feat: project post 수정 (글쓰기, 목표 및 참가자 연관관계 매핑 수정)

+++ src/main/java/com/matching/post/domain/QnaPost.java
@@ -24,8 +24,7 @@ public class QnaPost extends BaseEntity {
     private Long id;
     private String title;
     private String body;
-
-    private int likeCount;
+    private Long likeCount;
 
 
     @ManyToOne(fetch = FetchType.LAZY)
